vocab_range instead of vocab_size + fix add_special_tokens
